requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
watchdog==6.0.0
pypdf==5.0.1
langchain==0.3.3
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
python-docx==1.1.2
pydantic==2.9.2
python-dotenv==1.2.2
